Make reading of child process output more portable.
The code reading the output of the SDLMAME child process was
non-functional on OpenBSD because of differences of behavior from
poll(2). When poll(2) monitors output from a pipe which reaches EOF,
the revents flags returned by poll(2) differ between implementations.
Linux only sets the POLLHUP flag, while OpenBSD sets both the POLLIN and
the POLLHUP flags.
Since the condition variable given as argument to process_stdout_ready
is partially derived from the revents flags, process_stdout_ready on
OpenBSD is called with G_IO_IN|G_IO_HUP when we hit EOF. With these
flags, the event source is not removed, and process_stdout_ready is
called again and again.

Modify the process_read_line function to return the status reported by
g_io_channel_read_line. This makes it possible for process_stdout_ready
and process_stderr_ready to remove the event source when the status is
not G_IO_STATUS_NORMAL.