Index a GvaGameStore after we insert all the rows.
The index holds a mapping of ROM names to GtkTreeRowReferences.  If we
build the index as we're inserting rows, then row references have to be
updated each time we insert a new row.  Building the index after we're
done inserting rows yields a slight performance improvement, though not
as much as I was hoping.