Fix columns not being remembered.
As a GtkTreeView destroys its own columns during dispose() it emits
"columns-changed" signals.  We normally listen for that signal and save
the column names to GSettings, but obviously we don't want to do that
while the tree view is being disposed.  This was causing the "columns"
key in GSettings to be empty after every program exit.

With GTK2 I was checking the GtkTreeView's GTK_IN_DESTRUCTION flag to
know when to ignore the signals.  With that flag gone in GTK3 I had left
that bit of code commented out, not remembering what it was for.  This
uses a different approach to achieve the same thing -- though it's not
any cleaner than before, unfortunately.

I had the idea of writing a custom GtkTreeView subclass to deal with
this and to make it more of a self-contained widget, but the main tree
view is tangled up into *everything*.  Would need sorting out.  Too big
a job for now.