Do not monitor non-existent directories.
Do not create a GFileMonitor for a directory that doesn't exist.  This
doesn't seem to be an issue on Linux, but on OpenBSD at least the file
monitor trips immediately with a G_FILE_MONITOR_EVENT_DELETED, which
causes GVA to ask on startup if it should re-analyze ROMs.

What exposed this issue was the fact that the MAME package on OpenBSD
includes $HOME/.mame/roms in the default ROM path, but that directory
will not exist in a newly-created user account.

Related to GNOME bug #652465.