Do not treat start() as a GSourceFunc.
Idiotic mistake in main() where we schedule start() directly as a main
loop callback.  Problem is main loop callbacks must return a boolean,
and start(), which is only meant to run once, has a void return type.
So it's a coin toss whether start() will be run over and over again.