Bug 670102 - Crash when searching
In earlier versions of GVA we set the tree view search column to
DESCRIPTION every time we loaded a new model, since changing the model
sets the search column to -1.

However tree_view_search_equal() would still occasionally get the wrong
search column ID passed to it, which would trip a column ID assertion.
Never figured out why, or how to reliably reproduce it.

But since the search column for the type-ahead feature is always the
DESCRIPTION column, it finally dawned on me that we can just hard-code
the column ID in tree_view_search_equal() and do away with the assertion
and having to repeatedly set the search column -- just leave it unset.