audit: Give the audit window its own tree store
I want to convert GvaGameStore back to a list store.

The audit window was the reason I converted GvaGameStore to a tree store,
thinking how neat it would be to reuse the same model class for the main
window and audit window.  But I think it suffered a big performance hit
as a result.  It was a dumb reason in retrospect.